% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean Function}
\usage{
clean(df, normalize)
}
\arguments{
\item{df}{Dataframe for manipulation.}

\item{normalize}{TRUE to normalize temperature, salinity, dissolved oxygen mgl, and turbidity. FALSE to do nothing.}
}
\description{
Package to prepare data collected under NOAA's National
Estuarine Research Reserve System's (NERRS) National Monitoring
Program for analysis. Converts dates to a friendly format, renames
columns, and normalizes environmental variables from 0 to 1.
}
\examples{
df2 <- clean(df, T)

}
\keyword{NOAA}
\keyword{data-cleaning}
\keyword{water}
