% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean Function}
\usage{
clean(df, normalize)
}
\arguments{
\item{normalize}{Do you wish to normalize data? Defaults to TRUE}
}
\description{
Package to prepare data collected under NOAA's National
Estuarine Research Reserve System's (NERRS) National Monitoring
Program for analysis
}
\examples{
clean()

}
\keyword{water}
